+++ src/cykooz/buildout/fixnamespace/tests.py
@@ -57,6 +57,14 @@ def _create_fs_tree(root_dir: Path, paths: list[str]):
                 ],
                 ['root', 'root.sub2', 'root.sub2.sub4']
         ),
+        (
+                [
+                    'sub1/__init__.py',
+                    'sub2/sub3/__init__.py',
+                    'sub4/sub5/readme.txt',
+                ],
+                ['root', 'root.sub2']
+        ),
     ]
 )
 def test_fix_namespaces(tmp_path, paths, result):
@@ -74,6 +82,18 @@ def test_fix_namespaces(tmp_path, paths, result):
         f.write('scripts\n')
 
     fix_namespace_packages_txt(tmp_path, distinfo_name)
+    assert_namespace_packages(distinfo_dir, result)
+
+    # Without top_level.txt
+    os.remove(toplevel_file)
+    np_path = distinfo_dir / 'namespace_packages.txt'
+    if np_path.is_file():
+        os.remove(np_path)
+    fix_namespace_packages_txt(tmp_path, distinfo_name)
+    assert_namespace_packages(distinfo_dir, result)
+
+
+def assert_namespace_packages(distinfo_dir, result):
     namespace_packages_file = distinfo_dir / 'namespace_packages.txt'
     if result:
         assert namespace_packages_file.is_file()